SAMZA-1592: Fix Hello-Samza latest after Kafka 0.11 upgrade in Samza
After the Kafka-0.11 upgrade, the latest branch in hello-samza (and hence all our open-source tutorials that rely on it) are broken. Users who follow the tutorial will see a `NoSuchMethodError` when starting their job due to a mismatch of Kafka versions. This PR makes both Kafka versions consistent by updating the hello-samza tutorial to use Kafka-0.11

Author: Jagadish <jvenkatraman@linkedin.com>

Reviewers: Prateek M<pmaheshw@linkedin.com>

Closes #32 from vjagadish1989/latest-2